+++ tests/phases/design/test_semisubmersible_design.py
@@ -9,7 +9,11 @@ from itertools import product
 
 import pytest
 
-from ORBIT.phases.design import SemiSubmersibleDesign
+from ORBIT.core.library import extract_library_specs
+from ORBIT.phases.design import (
+    SemiSubmersibleDesign,
+    CustomSemiSubmersibleDesign,
+)
 
 base = {
     "site": {"depth": 500},
@@ -63,6 +67,144 @@ def test_design_kwargs():
         s = SemiSubmersibleDesign(config)
         s.run()
         cost = s.total_cost
+
+        assert cost != base_cost
+
+
+config_custom = extract_library_specs(
+    "config", "semisubmersible_design_custom"
+)
+
+
+def test_calc_geometric_scale_factor():
+
+    config = deepcopy(config_custom)
+
+    custom = CustomSemiSubmersibleDesign(config)
+    custom.run()
+
+    assert custom.geom_scale_factor == 1.0
+
+    config["turbine"]["rotor_diameter"] = 220
+    custom.run()
+
+    assert custom.geom_scale_factor == pytest.approx(0.9392, 1e-4)
+
+    with pytest.raises(KeyError):
+        config["turbine"]["rotor_diameter"] = None
+        custom.run()
+
+
+def test_bouyant_column_volume():
+
+    custom = CustomSemiSubmersibleDesign(config_custom)
+    custom.run()
+
+    assert custom.bouyant_column_volume == pytest.approx(72.5136, 1e-4)
+
+
+def test_center_column_volume():
+
+    custom = CustomSemiSubmersibleDesign(config_custom)
+    custom.run()
+
+    assert custom.center_column_volume == pytest.approx(39.4059, 1e-4)
+
+
+def test_pontoon_volume():
+
+    custom = CustomSemiSubmersibleDesign(config_custom)
+    custom.run()
+
+    assert custom.pontoon_volume == pytest.approx(90.4021, 1e-4)
+
+
+def test_strut_volume():
+
+    custom = CustomSemiSubmersibleDesign(config_custom)
+    custom.run()
+
+    assert custom.strut_volume == pytest.approx(32.9219, 1e-4)
+
+
+def test_substructure_steel_mass():
+
+    custom = CustomSemiSubmersibleDesign(config_custom)
+    custom.run()
+
+    assert custom.substructure_steel_mass == pytest.approx(5000, 1e0)
+
+
+def test_ballast_mass():
+
+    custom = CustomSemiSubmersibleDesign(config_custom)
+    custom.run()
+
+    assert custom.ballast_mass == 2540
+
+
+def test_tower_interface_mass():
+
+    custom = CustomSemiSubmersibleDesign(config_custom)
+    custom.run()
+
+    assert custom.tower_interface_mass == 100
+
+
+def test_substructure_steel_cost():
+
+    custom = CustomSemiSubmersibleDesign(config_custom)
+    custom.run()
+
+    assert custom.steel_cr == 4500
+
+
+def test_substructure_mass():
+
+    custom = CustomSemiSubmersibleDesign(config_custom)
+    custom.run()
+
+    assert custom.substructure_mass == pytest.approx(7642, 1e0)
+
+
+def test_substructure_unit_cost():
+
+    custom = CustomSemiSubmersibleDesign(config_custom)
+    custom.run()
+
+    assert custom.substructure_unit_cost == pytest.approx(2.3343e7, 1e-4)
+
+
+def test_custom_design_kwargs():
+
+    test_kwargs = {
+        "column_diameter": 15,
+        "wall_thickness": 0.1,
+        "column_height": 20,
+        "pontoon_length": 52,
+        "pontoon_width": 15,
+        "pontoon_height": 5,
+        "strut_diameter": 1.0,
+        "steel_density": 8500,
+        "ballast_mass": 1000,
+        "tower_interface_mass": 125,
+        "steel_CR": 3000,
+        "ballast_material_CR": 200,
+    }
+
+    cust = CustomSemiSubmersibleDesign(config_custom)
+    cust.run()
+    base_cost = cust.total_cost
+
+    for k, v in test_kwargs.items():
+
+        config = deepcopy(config_custom)
+        config["semisubmersible_design"] = {}
+        config["semisubmersible_design"][k] = v
+
+        cust = CustomSemiSubmersibleDesign(config)
+        cust.run()
+        cost = cust.total_cost
 
         assert cost != base_cost